+++ imod/msw/utilities/imod5_converter.py
@@ -1,3 +1,8 @@
+from typing import TypedDict, cast
+
+import numpy as np
+import pandas as pd
+import xarray as xr
 from xarray.core.utils import is_scalar
 
 from imod.common.constants import MaskValues
@@ -5,9 +10,29 @@ from imod.logging import LogLevel, logger
 from imod.mf6 import StructuredDiscretization
 from imod.msw.utilities.common import concat_imod5
 from imod.msw.utilities.mask import MetaSwapActive
-from imod.typing import GridDataArray, GridDataDict
-from imod.typing.grid import ones_like
+from imod.typing import GridDataArray, GridDataDict, Imod5DataDict
+from imod.typing.grid import ones_like, zeros_like
 from imod.util.spatial import get_cell_area
+
+
+# Some additional type aliases for sprinkling data, which is a bit more complex
+# than other packages.
+class CapSprinklingDataDict(TypedDict, total=False):
+    artificial_recharge: GridDataArray
+    artificial_recharge_layer: pd.DataFrame
+    artificial_recharge_capacity: GridDataArray
+
+
+class SprinklingPointsDataDict(TypedDict, total=False):
+    x_p: np.ndarray | list[float]
+    y_p: np.ndarray | list[float]
+    layer_p: np.ndarray | list[int]
+    id_sprinkling_p: np.ndarray | list[int]
+    capacity_p: np.ndarray | list[float]
+
+
+class SprinklingPointsGridDataDict(SprinklingPointsDataDict, total=False):
+    art_grid: GridDataArray
 
 
 def get_cell_area_from_imod5_data(
@@ -116,3 +141,85 @@ def has_active_scaling_factor(imod5_cap: GridDataDict):
         )
 
     return not scaling_factor_inactive
+
+
+def is_sprinkling_from_points(imod5_data: Imod5DataDict) -> bool:
+    """
+    Check if sprinkling is specified from points, based on the presence of
+    sprinkling layer and sprinkling points data in the iMOD5 CAP dataset.
+    """
+    cap_data = cast(CapSprinklingDataDict, imod5_data["cap"])
+    if isinstance(cap_data.get("artificial_recharge_layer"), pd.DataFrame):
+        return True
+    return False
+
+
+def sprinkling_data_from_imod5_ipf(
+    cap_data: CapSprinklingDataDict,
+) -> SprinklingPointsGridDataDict:
+    art_grid = cap_data["artificial_recharge"]
+    # Set urban landuse irrigation to 0, as sprinkling is not allowed for urban landuse.
+    subunit_template = xr.DataArray(
+        np.array([1, 0], dtype=int), dims="subunit", coords={"subunit": [0, 1]}
+    )
+    art_grid = subunit_template * art_grid
+
+    df_points = cap_data["artificial_recharge_layer"]
+    # Select first 5 columns and enforce column names, iMOD5 expects columns in
+    # this order. The additional columns are metadata for the user and can be
+    # ignored.
+    arl_points = df_points.iloc[:, :5]
+    arl_points.columns = ["x_p", "y_p", "layer_p", "id_sprinkling_p", "capacity_p"]
+    # Enforce dtypes
+    dtype_dict = {
+        "x_p": float,
+        "y_p": float,
+        "layer_p": int,
+        "id_sprinkling_p": int,
+        "capacity_p": float,
+    }
+
+    arl_points = arl_points.astype(dtype_dict)
+    arl_point_dict = cast(
+        SprinklingPointsDataDict,
+        {key: arl_points[key].to_numpy() for key in dtype_dict.keys()},
+    )
+
+    return {
+        "art_grid": art_grid,
+        **arl_point_dict,
+    }
+
+
+def sprinkling_data_from_imod5_grid(cap_data: GridDataDict) -> GridDataDict:
+    # Convert units from mm/d to m3/d
+    msw_area = get_cell_area_from_imod5_data(cap_data)
+    capacity_mmd = cap_data["artificial_recharge_capacity"]
+    capacity_m3d = capacity_mmd * 1e-3 * msw_area.sel(subunit=0, drop=True)
+
+    artificial_rch_type = cap_data["artificial_recharge"]
+    from_groundwater = artificial_rch_type == 1
+    from_surfacewater = artificial_rch_type == 2
+    is_active = artificial_rch_type != 0
+
+    zero_where_active = zeros_like(artificial_rch_type).where(is_active)
+
+    # Add zero where active, to have active cells set to 0.0.
+    max_abstraction_groundwater_rural = zero_where_active.where(
+        ~from_groundwater, capacity_m3d
+    )
+    max_abstraction_surfacewater_rural = zero_where_active.where(
+        ~from_surfacewater, capacity_m3d
+    )
+
+    # No sprinkling for urban environments
+    max_abstraction_urban = zero_where_active
+
+    data = {}
+    data["max_abstraction_groundwater"] = concat_imod5(
+        max_abstraction_groundwater_rural, max_abstraction_urban
+    )
+    data["max_abstraction_surfacewater"] = concat_imod5(
+        max_abstraction_surfacewater_rural, max_abstraction_urban
+    )
+    return data
